fix amd
